# carbon: use a generator in routers instead of returning the full set