- This preserve the order
- This stops iterating as soon as we have enough nodes (accoring
  to replication factor).

This restore pre-1.1.0 performances (/10 CPU usage).

Benchmarks follow .. (we should really move to pytest to have better output).

Before:
```
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 36
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 456
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 761
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.28447
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 35
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 540
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 884
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.26407
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 33
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 544
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 871
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.34426
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 31
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 489
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 836
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.32029
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 32
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 541
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 840
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.34689
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 30
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 452
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 813
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.2657
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 31
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 543
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 839
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.31652
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 33
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 504
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 905
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.27626
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 30
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 340
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 615
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 msecs: 958
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 27
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 309
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 600
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 msecs: 937
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 29
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 343
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 677
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 secs: 1.06069
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 26
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 340
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 632
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 msecs: 967
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 29
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 550
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 902
 consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.26543
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 25
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 543
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 914
 consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.25108
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 29
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 587
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 955
 consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.34788
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 25
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 565
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 939
 consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.3025
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 84
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 467
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 804
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.27919
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 54
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 499
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 798
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.22153
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 55
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 632
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 927
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.38
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 52
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 572
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 898
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: None datapoints: 100000 secs: 1.34593
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 58
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 485
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 815
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.28422
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 51
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 513
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 839
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.27546
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 59
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 576
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 895
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.35766
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: carbon_ch datapoints: 100000 msecs: 49
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: carbon_ch datapoints: 100000 msecs: 584
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: carbon_ch datapoints: 100000 msecs: 910
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: carbon_ch datapoints: 100000 secs: 1.36756
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 56
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 331
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 633
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 msecs: 979
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 48
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 326
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 623
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 secs: 1.00448
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 63
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 399
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 713
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 secs: 1.07203
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: fnv1a_ch datapoints: 100000 msecs: 47
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: fnv1a_ch datapoints: 100000 msecs: 370
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: fnv1a_ch datapoints: 100000 msecs: 718
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: fnv1a_ch datapoints: 100000 secs: 1.04931
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 49
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 616
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 967
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.30962
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 48
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 610
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 msecs: 946
 aggregated-consistent-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.3382
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 54
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 667
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 secs: 1.03899
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.43022
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: mmh3_ch datapoints: 100000 msecs: 49
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 16    hash_type: mmh3_ch datapoints: 100000 msecs: 649
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 32    hash_type: mmh3_ch datapoints: 100000 secs: 1.06055
 aggregated-consistent-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 48    hash_type: mmh3_ch datapoints: 100000 secs: 1.39583
 fast-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 246
 fast-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 236
 fast-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 259
 fast-hashing  replication_factor: 1 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 msecs: 238
 fast-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 239
 fast-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 241
 fast-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 240
 fast-hashing  replication_factor: 1 diverse_replicas: 0 n_destinations: 48    hash_type: None datapoints: 100000 msecs: 253
 fast-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 240
 fast-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 16    hash_type: None datapoints: 100000 msecs: 298
 fast-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 32    hash_type: None datapoints: 100000 msecs: 296
 fast-hashing  replication_factor: 4 diverse_replicas: 1 n_destinations: 48    hash_type: None datapoints: 100000 msecs: 298
 fast-hashing  replication_factor: 4 diverse_replicas: 0 n_destinations: 1     hash_type: None datapoints: 100000 msecs: 244
…